Phase 2 step 11: Migrate TradingEngineMixin + AgentChatMixin (co-located engine/agent callbacks, lazy fallback imports)
