+++ apps/tui/dashboard_tui.py
@@ -1020,15 +1020,6 @@ class NepseDashboard(
         if self.trade_mode == "paper":
             self._start_all_account_engines()
 
-    @work(thread=True)
-    def _init_tms(self) -> None:
-        """Live brokerage not available in public release."""
-        pass
-
-    def _attach_tms_live_source(self) -> None:
-        """Live brokerage not available in public release."""
-        pass
-
     # ── News ticker ──────────────────────────────────────────────────────────
 
     def _build_ticker(self) -> None:
@@ -1291,292 +1282,12 @@ class NepseDashboard(
         except Exception:
             pass
 
-    # ── Trading Engine ─────────────────────────────────────────────────────────
-
-    @work(thread=True)
-    def _start_trading_loop(self) -> None:
-        """Run the auto-trading engine in a background thread (legacy single-account)."""
-        if self._trading_engine:
-            self._trading_engine.run_loop()
-
-    # ── Multi-account engine management ────────────────────────────────────────
-
-    def _start_all_account_engines(self) -> None:
-        """Create and start one TUITradingEngine per account."""
-        self._account_engines = {}
-        for account in list(getattr(self, "_paper_accounts", []) or []):
-            account_id = str(account.get("id") or "")
-            if not account_id:
-                continue
-            strategy_id = str(account.get("strategy_id") or "")
-            strategy = strategy_registry.load_strategy(strategy_id) if strategy_id else None
-            config = dict((strategy or {}).get("config") or {}) or dict(LONG_TERM_CONFIG)
-            account_dir = _account_dir(account_id)
-            account_capital = _account_initial_capital_from_files(
-                account_dir,
-                float(config.get("initial_capital") or INITIAL_CAPITAL),
-            )
-            engine = TUITradingEngine(
-                capital=account_capital,
-                signal_types=list(config.get("signal_types") or list(LONG_TERM_CONFIG.get("signal_types") or [])),
-                max_positions=int(config.get("max_positions") or LONG_TERM_CONFIG.get("max_positions") or 5),
-                holding_days=int(config.get("holding_days") or LONG_TERM_CONFIG.get("holding_days") or 40),
-                sector_limit=float(config.get("sector_limit") or LONG_TERM_CONFIG.get("sector_limit") or 0.35),
-                stop_loss_pct=float(config.get("stop_loss_pct") or LONG_TERM_CONFIG.get("stop_loss_pct") or 0.08),
-                trailing_stop_pct=float(config.get("trailing_stop_pct") or LONG_TERM_CONFIG.get("trailing_stop_pct") or 0.10),
-                hedge_enabled=bool(getattr(self, "_hedge_enabled", True)),
-                portfolio_file=account_dir / "paper_portfolio.csv",
-                trade_log_file=account_dir / "paper_trade_log.csv",
-                nav_log_file=account_dir / "paper_nav_log.csv",
-                state_file=account_dir / "paper_state.json",
-                account_id=account_id,
-                strategy_id=strategy_id,
-                strategy_config=config,
-                on_status=self._make_engine_status_cb(account_id),
-                on_activity=self._make_engine_activity_cb(account_id),
-                on_portfolio_changed=self._make_engine_portfolio_cb(account_id),
-                on_agent_updated=self._make_engine_agent_cb(account_id),
-            )
-            self._account_engines[account_id] = engine
-            if account_id == self._current_account_id:
-                self._trading_engine = engine
-            self._start_account_engine_worker(account_id)
-
-    @work(thread=True)
-    def _start_account_engine_worker(self, account_id: str) -> None:
-        """Run one account's trading engine in a dedicated background worker thread."""
-        engine = self._account_engines.get(account_id)
-        if engine:
-            engine.run_loop()
-
-    def _make_engine_status_cb(self, account_id: str):
-        """Status callback — only updates UI when this account is active."""
-        def cb(msg: str) -> None:
-            if account_id == getattr(self, "_current_account_id", ""):
-                self.call_from_thread(self._set_status, msg)
-        return cb
-
-    def _make_engine_activity_cb(self, account_id: str):
-        """Activity log callback — only updates UI when this account is active."""
-        def cb(msg: str) -> None:
-            if account_id == getattr(self, "_current_account_id", ""):
-                self.call_from_thread(self._append_activity, msg)
-        return cb
-
-    def _make_engine_portfolio_cb(self, account_id: str):
-        """Portfolio-changed callback — only refreshes UI when this account is active."""
-        def cb() -> None:
-            if account_id == getattr(self, "_current_account_id", ""):
-                self.call_from_thread(self._on_engine_portfolio_changed)
-        return cb
-
-    def _make_engine_agent_cb(self, account_id: str):
-        """Agent-updated callback — only refreshes UI when this account is active."""
-        def cb() -> None:
-            if account_id == getattr(self, "_current_account_id", ""):
-                self.call_from_thread(self._on_engine_agent_updated)
-        return cb
-
-    def _append_activity(self, msg: str) -> None:
-        """Append a message to the activity log in the portfolio tab."""
-        try:
-            scroll = self.query_one("#activity-scroll", VerticalScroll)
-            label = Static(msg, classes="activity-line")
-            scroll.mount(label)
-            scroll.scroll_end(animate=False)
-            # Keep max 100 lines
-            children = list(scroll.children)
-            if len(children) > 100:
-                for child in children[:len(children) - 100]:
-                    child.remove()
-        except Exception:
-            pass
-
-    def _on_engine_portfolio_changed(self) -> None:
-        """Called by engine when positions change — refresh portfolio display."""
-        if self._trading_engine:
-            self._stats = self._trading_engine.get_portfolio_stats()
-            self._populate_portfolio_tab(self._stats)
-            self._populate_risk_tab(self._stats)
-            self._populate_trades_full()
-        self._render_hedge_panel()
-
-    def _on_engine_agent_updated(self) -> None:
-        """Called by engine after agent analysis — refresh agents tab."""
-        try:
-            self._load_agent_runtime_state()
-            self._populate_agent_tab()
-        except Exception:
-            pass
-
-    def _load_agent_runtime_state(self) -> None:
-        self._agent_analysis = load_agent_analysis() or {}
-        current_account_id = str(getattr(self, "_current_account_id", "account_1") or "account_1")
-        current_strategy_id = str((self._strategy_account_binding(current_account_id) or {}).get("strategy_id") or strategy_registry.default_strategy_for_account(current_account_id))
-        if (
-            str((self._agent_analysis or {}).get("account_id") or "") != current_account_id
-            or str((self._agent_analysis or {}).get("strategy_id") or "") != current_strategy_id
-        ):
-            self._agent_analysis = {}
-        all_recent = list(load_agent_history() or [])
-        visible_cutoff = float(getattr(self, "_agent_visible_since", 0.0) or 0.0)
-        self._agent_history, self._agent_hidden_recent_history = _split_agent_messages_by_cutoff(all_recent, visible_cutoff)
-        archived_items = list(load_agent_archive_history() or [])
-        self._agent_archive_count = len(archived_items) + len(self._agent_hidden_recent_history)
-        if self._agent_show_archived:
-            older = archived_items + list(self._agent_hidden_recent_history or [])
-            self._agent_archived_history = older[-AGENT_ARCHIVE_RENDER_LIMIT:]
-        else:
-            self._agent_archived_history = []
-
     def _active_account_name(self) -> str:
         account_id = str(getattr(self, "_current_account_id", "account_1") or "account_1")
         for account in list(getattr(self, "_paper_accounts", []) or []):
             if str(account.get("id") or "") == account_id:
                 return str(account.get("name") or account_id)
         return account_id
-
-    def _sync_agent_account_context_env(self) -> None:
-        account_id = str(getattr(self, "_current_account_id", "account_1") or "account_1")
-        account_dir = _account_dir(account_id)
-        strategy_id = str((self._strategy_account_binding(account_id) or {}).get("strategy_id") or strategy_registry.default_strategy_for_account(account_id))
-        os.environ["NEPSE_ACTIVE_ACCOUNT_ID"] = account_id
-        os.environ["NEPSE_ACTIVE_ACCOUNT_NAME"] = self._active_account_name()
-        os.environ["NEPSE_ACTIVE_ACCOUNT_DIR"] = str(account_dir)
-        os.environ["NEPSE_ACTIVE_PORTFOLIO_FILE"] = str(account_dir / "paper_portfolio.csv")
-        os.environ["NEPSE_ACTIVE_STRATEGY_ID"] = strategy_id
-        os.environ["NEPSE_ACTIVE_STRATEGY_NAME"] = strategy_registry.strategy_name(strategy_id)
-
-    def _current_agent_provider_label(self) -> str:
-        try:
-            cfg = load_active_agent_config()
-        except Exception:
-            cfg = {}
-        configured = str(
-            cfg.get("provider_label")
-            or cfg.get("backend")
-            or os.environ.get("NEPSE_AGENT_PROVIDER_LABEL")
-            or "gemma4_mlx"
-        ).strip()
-        if configured:
-            return configured
-        meta = dict((getattr(self, "_agent_analysis", {}) or {}).get("agent_runtime_meta") or {})
-        return str(meta.get("provider") or "gemma4_mlx")
-
-    def _agent_runtime_summary(self) -> str:
-        cfg = load_active_agent_config()
-        preset = str(cfg.get("selected_preset") or cfg.get("backend") or "ollama")
-        backend = str(cfg.get("backend") or preset)
-        model = str(cfg.get("model") or "").strip() or "default"
-        return f"{preset} backend={backend} model={model}"
-
-    def _agent_backends_help(self) -> str:
-        backends = list_agent_backends()
-        parts = [
-            f"{row['id']} ({row.get('model') or 'no default model'})"
-            for row in backends
-        ]
-        return "Agent backends: " + ", ".join(parts)
-
-    def _stop_active_agent_chat(self, *, announce: bool = True) -> bool:
-        proc = getattr(self, "_agent_chat_process", None)
-        if proc is None:
-            return False
-        self._agent_chat_stop_requested = True
-        try:
-            proc.terminate()
-        except Exception:
-            pass
-        try:
-            proc.wait(timeout=2)
-        except Exception:
-            try:
-                proc.kill()
-            except Exception:
-                pass
-        self._agent_chat_process = None
-        self.call_from_thread(self._set_typing, False)
-        if announce:
-            provider = self._current_agent_provider_label()
-            append_external_agent_chat_message("AGENT", "Stopped by user.", source="tui_chat", provider=provider)
-            self.call_from_thread(self._append_chat, "AGENT", "Stopped by user.", PURPLE)
-            self.call_from_thread(self._set_status, "Agent response stopped")
-        return True
-
-    def _update_agent_chat_hint(self) -> None:
-        hint = self.query_one("#agent-chat-hint", Static)
-        if getattr(self, "_agent_show_archived", False):
-            text = "Archive view loaded. Type /recent to collapse."
-        else:
-            text = ""
-        hint.styles.height = 1 if text else 0
-        hint.update(Text(text, style=LABEL))
-
-    def _get_agent_typing_widget(self) -> Static:
-        widget = getattr(self, "_agent_typing_widget", None)
-        if widget is None:
-            widget = Static("", classes="chat-msg chat-note", id="agent-typing-line")
-            self._agent_typing_widget = widget
-        return widget
-
-    def _detach_agent_typing_widget(self) -> None:
-        widget = getattr(self, "_agent_typing_widget", None)
-        if widget is not None and widget.parent is not None:
-            widget.remove()
-
-    def _ensure_agent_typing_widget(self) -> Static:
-        scroll = self.query_one("#agent-chat-scroll", VerticalScroll)
-        widget = self._get_agent_typing_widget()
-        if widget.parent is None:
-            scroll.mount(widget)
-        elif widget.parent is not scroll:
-            widget.remove()
-            scroll.mount(widget)
-        return widget
-
-    def _append_chat_note(self, message: str) -> None:
-        scroll = self.query_one("#agent-chat-scroll", VerticalScroll)
-        note_text = Text.assemble(
-            ("[SYSTEM]: ", f"bold {LABEL}"),
-            (str(message or "").strip(), LABEL),
-        )
-        widget = Static(note_text, classes="chat-msg chat-note")
-        scroll.mount(widget)
-        scroll.scroll_end(animate=False)
-
-    def _render_agent_chat_history(self) -> None:
-        scroll = self.query_one("#agent-chat-scroll", VerticalScroll)
-        scroll.remove_children()
-        archived = list(getattr(self, "_agent_archived_history", []) or [])
-        recent = list(getattr(self, "_agent_history", []) or [])
-        if getattr(self, "_agent_show_archived", False) and archived:
-            self._append_chat_note(f"ARCHIVE · showing {len(archived)} older messages")
-            for item in archived:
-                role = str(item.get("role") or "AGENT").upper()
-                color = BLUE if role == "YOU" else AMBER if role == "AGENT" else LABEL
-                self._append_chat(
-                    role,
-                    str(item.get("message") or ""),
-                    color,
-                    ts=item.get("ts"),
-                    provider=item.get("provider"),
-                )
-            self._append_chat_note("RECENT")
-        for item in recent:
-            role = str(item.get("role") or "AGENT").upper()
-            color = BLUE if role == "YOU" else AMBER if role == "AGENT" else LABEL
-            self._append_chat(
-                role,
-                str(item.get("message") or ""),
-                color,
-                ts=item.get("ts"),
-                provider=item.get("provider"),
-            )
-        if bool(getattr(self, "_agent_typing_visible", False)):
-            self._animate_agent_typing()
-        else:
-            self._detach_agent_typing_widget()
-        self._update_agent_chat_hint()
 
     # ── Auto-refresh ──────────────────────────────────────────────────────────
 
@@ -2343,344 +2054,6 @@ class NepseDashboard(
 
     # ── Agent tab ─────────────────────────────────────────────────────────────
 
-    @work(thread=True)
-    def _run_agent_analysis(self, force: bool = True) -> None:
-        self._sync_agent_account_context_env()
-        self.call_from_thread(self._set_status, "⧖ Agent analyzing..." if force else "⧖ Loading agent...")
-        self.call_from_thread(self._update_agent_status, "ANALYZING..." if force else "LOADING...", "running")
-        try:
-            preview = build_algo_shortlist_snapshot()
-            self.call_from_thread(self._set_agent_shortlist_preview, preview)
-            result = agent_analyze(force=force)
-            self._agent_analysis = result
-            self.call_from_thread(self._populate_agent_tab)
-            self.call_from_thread(self._maybe_submit_agent_super_signal, result)
-            age = time.time() - result.get("timestamp", 0)
-            if age < 5:
-                self.call_from_thread(self._set_status,
-                    f"Agent analysis complete │ {len(result.get('stocks', []))} stocks reviewed")
-            else:
-                self.call_from_thread(self._set_status,
-                    f"Agent loaded from cache ({int(age/60)}m ago) │ Press A+Enter to refresh")
-        except Exception as e:
-            self.call_from_thread(self._set_status, f"Agent error: {e}")
-
-    def _set_agent_shortlist_preview(self, preview: dict) -> None:
-        current_rows = list((getattr(self, "_agent_analysis", {}) or {}).get("stocks") or [])
-        if current_rows and not all(str(row.get("verdict") or "").upper() == "REVIEW" for row in current_rows):
-            return
-        self._agent_preview_override = dict(preview or {})
-        self._populate_agent_tab()
-
-    def _set_typing(self, visible: bool):
-        self._agent_typing_visible = bool(visible)
-        if not visible:
-            self._agent_typing_frame = 0
-            self._detach_agent_typing_widget()
-        else:
-            self._animate_agent_typing()
-
-    def _animate_agent_typing(self) -> None:
-        if not bool(getattr(self, "_agent_typing_visible", False)):
-            self._detach_agent_typing_widget()
-            return
-        typing_w = self._ensure_agent_typing_widget()
-        frames = ["[AGENT] : Typing.", "[AGENT] : Typing..", "[AGENT] : Typing..."]
-        frame = frames[int(getattr(self, "_agent_typing_frame", 0)) % len(frames)]
-        self._agent_typing_frame = int(getattr(self, "_agent_typing_frame", 0)) + 1
-        typing_w.update(Text(frame, style=f"italic {LABEL}"))
-        try:
-            self.query_one("#agent-chat-scroll", VerticalScroll).scroll_end(animate=False)
-        except Exception:
-            pass
-
-    @work(thread=True)
-    def _agent_ask_async(self, question: str) -> None:
-        existing = getattr(self, "_agent_chat_process", None)
-        if existing is not None and existing.poll() is None:
-            self.call_from_thread(self._set_status, "Agent is already thinking | use /stop to cancel")
-            return
-
-        self._sync_agent_account_context_env()
-        self._agent_chat_request_id = int(getattr(self, "_agent_chat_request_id", 0)) + 1
-        request_id = self._agent_chat_request_id
-        self._agent_chat_stop_requested = False
-        provider = self._current_agent_provider_label()
-        append_external_agent_chat_message("YOU", question, source="tui_chat", provider=provider)
-        self.call_from_thread(self._append_chat, "YOU", question, BLUE)
-        try:
-            cmd = [
-                sys.executable,
-                str(PROJECT_ROOT / "scripts" / "agents" / "run_active_agent.py"),
-                "--question",
-                question,
-            ]
-            env = dict(os.environ)
-            env["NEPSE_AGENT_DISABLE_HISTORY"] = "1"
-            proc = subprocess.Popen(
-                cmd,
-                cwd=str(PROJECT_ROOT),
-                env=env,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-            )
-            self._agent_chat_process = proc
-            self.call_from_thread(self._set_typing, True)
-            self.call_from_thread(self._set_status, "⧖ Agent thinking...")
-
-            started = time.monotonic()
-            while True:
-                if request_id != int(getattr(self, "_agent_chat_request_id", 0)):
-                    self._stop_active_agent_chat(announce=False)
-                    return
-                if bool(getattr(self, "_agent_chat_stop_requested", False)):
-                    self._stop_active_agent_chat(announce=False)
-                    return
-                if proc.poll() is not None:
-                    break
-                if (time.monotonic() - started) > AGENT_CHAT_TIMEOUT_SECS:
-                    self._stop_active_agent_chat(announce=False)
-                    append_external_agent_chat_message(
-                        "AGENT",
-                        f"Timed out after {AGENT_CHAT_TIMEOUT_SECS}s. Ask a tighter question or try again.",
-                        source="tui_chat",
-                        provider=provider,
-                    )
-                    self.call_from_thread(self._set_typing, False)
-                    self.call_from_thread(
-                        self._append_chat,
-                        "AGENT",
-                        f"Timed out after {AGENT_CHAT_TIMEOUT_SECS}s. Ask a tighter question or try again.",
-                        PURPLE,
-                    )
-                    self.call_from_thread(self._set_status, "Agent timed out")
-                    return
-                time.sleep(0.1)
-
-            stdout, stderr = proc.communicate(timeout=1)
-            self._agent_chat_process = None
-            answer = str(stdout or "").strip()
-            if not answer:
-                answer = " ".join(str(stderr or "Agent returned no output.").split())[:280]
-            append_external_agent_chat_message("AGENT", answer, source="tui_chat", provider=provider)
-            self.call_from_thread(self._set_typing, False)
-            self.call_from_thread(self._append_chat, "AGENT", answer, PURPLE)
-            self.call_from_thread(self._set_status, "Agent responded")
-        except Exception as e:
-            self._agent_chat_process = None
-            self.call_from_thread(self._set_typing, False)
-            self.call_from_thread(self._append_chat, "ERROR", str(e), LOSS_HI)
-
-    def _update_agent_status(self, text_str: str, state: str = "idle"):
-        bar = self.query_one("#agent-status-bar", Static)
-        a = getattr(self, '_agent_analysis', {})
-        if state == "running":
-            bar.update(Text.from_markup(
-                f"[bold #bb88ff]◆ AGENT[/]   [{YELLOW}]⧖ {text_str}[/]"))
-            return
-        trade = a.get("trade_today", None)
-        tc = GAIN_HI if trade else LOSS_HI if trade is False else LABEL
-        trade_str = "YES" if trade else "NO" if trade is False else "?"
-        stocks = list(a.get("stocks", []) or a.get("shortlist", []) or [])
-        n_approve = sum(1 for s in stocks if s.get("verdict", "").upper() == "APPROVE")
-        n_reject = sum(1 for s in stocks if s.get("verdict", "").upper() == "REJECT")
-        n_total = len(stocks)
-        n_super = sum(1 for s in stocks if bool(s.get("auto_entry_candidate")))
-        regime = a.get("regime", "?")
-        meta = dict(a.get("agent_runtime_meta") or {})
-        provider = self._current_agent_provider_label().upper()
-        import time as _t
-        age = _t.time() - a.get("timestamp", 0)
-        age_str = f"{int(age/60)}m ago" if age < 3600 else f"{int(age/3600)}h ago" if age < 86400 else "stale"
-        bar.update(Text.from_markup(
-            f"[bold #bb88ff]◆ {provider} AGENT[/]   "
-            f"[#888888]Trade today:[/] [bold {tc}]{trade_str}[/]   "
-            f"[#888888]Verdicts:[/] [bold {GAIN_HI}]{n_approve}✓[/] "
-            f"[bold {LOSS_HI}]{n_reject}✗[/] "
-            f"[#888888]of {n_total}[/]   "
-            f"[#888888]Super:[/] [bold {AMBER}]{n_super}[/]   "
-            f"[#888888]Regime:[/] [{YELLOW}]{regime}[/]   "
-            f"[#888888]Updated:[/] [{LABEL}]{age_str}[/]   "
-            f"[#555555]A=Analyze  │  /agent  │  /model  │  /history[/]"
-        ))
-
-    def _populate_agent_meta_headers(self, analysis: dict) -> None:
-        stocks = list((analysis or {}).get("stocks") or [])
-        left = self.query_one("#agent-picks-subtitle", Static)
-        right = self.query_one("#agent-chat-subtitle", Static)
-        provider = self._current_agent_provider_label().upper()
-        if stocks:
-            buy_count = sum(1 for row in stocks if str(row.get("action_label") or "").upper() == "BUY")
-            super_count = sum(1 for row in stocks if bool(row.get("auto_entry_candidate")))
-            left.update(
-                Text(
-                    f"Ranked shortlist · {len(stocks)} names · {buy_count} buys · {super_count} super signals",
-                    style=LABEL,
-                )
-            )
-        else:
-            left.update(Text("Refreshing live snapshot and building the ranked shortlist.", style=LABEL))
-        archive_count = int(getattr(self, "_agent_archive_count", 0) or 0)
-        right.update(
-            Text(
-                f"{provider} conversation stream · live session",
-                style=LABEL,
-            )
-        )
-
-    def _populate_agent_market_banner(self, analysis: dict) -> None:
-        banner = self.query_one("#agent-market-view", Static)
-        stocks = list((analysis or {}).get("stocks") or [])
-        fresh_market = dict((analysis or {}).get("fresh_market") or {})
-        parts: list[str] = []
-        regime = str((analysis or {}).get("regime") or "unknown").upper()
-        if regime and regime != "UNKNOWN":
-            parts.append(f"[#888888]Regime[/] [bold {YELLOW}]{regime}[/]")
-        session_date = str(fresh_market.get("session_date") or (analysis or {}).get("context_date") or "")
-        if session_date:
-            parts.append(f"[#888888]Session[/] [bold {WHITE}]{session_date}[/]")
-        quote_count = int(fresh_market.get("quote_count") or 0)
-        if quote_count > 0:
-            parts.append(f"[#888888]Quotes[/] [bold {AMBER}]{quote_count}[/]")
-        if stocks:
-            top = stocks[0]
-            parts.append(
-                f"[#888888]Top[/] [bold {AMBER}]{str(top.get('symbol') or '—')}[/] "
-                f"[{WHITE}]{str(top.get('action_label') or top.get('verdict') or 'REVIEW').upper()}[/] "
-                f"[#888888]score[/] [bold {AMBER}]{float(top.get('signal_score') or 0.0):.2f}[/]"
-            )
-        elif not parts:
-            parts.append("[#888888]Agent snapshot is loading[/]")
-        banner.update(Text.from_markup("   [#444444]│[/]   ".join(parts)))
-
-    def _populate_agent_detail_default(self, analysis: dict) -> None:
-        try:
-            self.query_one("#agent-detail-title", Static).update("FOCUS")
-        except Exception:
-            pass
-        detail = self.query_one("#agent-detail", Static)
-        stocks = list((analysis or {}).get("stocks") or [])
-        fresh_market = dict((analysis or {}).get("fresh_market") or {})
-        if stocks:
-            top = stocks[0]
-            top_symbol = str(top.get("symbol") or "—")
-            top_action = str(top.get("action_label") or top.get("verdict") or "REVIEW").upper()
-            top_score = float(top.get("signal_score") or 0.0)
-            top_conv = float(top.get("conviction") or 0.0)
-            summary = (
-                f"Top setup: {top_symbol} {top_action}  "
-                f"score {top_score:.2f}  conv {top_conv:.0%}\n"
-                f"{str(top.get('what_matters') or top.get('reasoning') or '')[:220]}"
-            )
-            detail.update(Text(summary, style=WHITE))
-            return
-        session_date = str(fresh_market.get("session_date") or (analysis or {}).get("context_date") or "—")
-        source = str(fresh_market.get("source") or "snapshot")
-        quote_count = int(fresh_market.get("quote_count") or 0)
-        detail.update(
-            Text(
-                f"Waiting for ranked picks.\n"
-                f"Session: {session_date}  Source: {source}  Quotes: {quote_count}",
-                style=LABEL,
-            )
-        )
-
-    def _show_agent_focus_row(self, index: int) -> None:
-        analysis = getattr(self, "_agent_analysis", {}) or {}
-        stocks = list(analysis.get("stocks") or [])
-        if not (0 <= int(index) < len(stocks)):
-            self._populate_agent_detail_default(analysis)
-            return
-
-        stock = dict(stocks[int(index)] or {})
-        symbol = str(stock.get("symbol") or "").upper()
-        verdict = str(stock.get("verdict") or "?").upper()
-        action_label = str(stock.get("action_label") or verdict or "REVIEW").upper()
-        action_color = {
-            "BUY": GAIN_HI,
-            "SELL": LOSS_HI,
-            "PASS": LOSS_HI,
-            "HOLD": YELLOW,
-            "REVIEW": LABEL,
-        }.get(action_label, WHITE)
-        conviction = float(stock.get("conviction", 0) or 0.0)
-        signal_score = float(stock.get("signal_score") or 0.0)
-        detail = Text.assemble(
-            (f"[{symbol}] ", f"bold {AMBER}"),
-            (f"{action_label} ", f"bold {action_color}"),
-            (f"score {signal_score:.2f} ", YELLOW),
-            (f"conv {conviction:.0%}\n", YELLOW),
-            (str(stock.get("what_matters") or stock.get("reasoning") or ""), WHITE),
-            ("\n", WHITE),
-            (f"Bull: {str(stock.get('bull_case') or 'n/a')}\n", GAIN),
-            (f"Risk: {str(stock.get('bear_case') or 'n/a')}", LOSS if stock.get("bear_case") else LABEL),
-        )
-        self.query_one("#agent-detail-title", Static).update(f"FOCUS · {symbol}")
-        self.query_one("#agent-detail", Static).update(detail)
-
-    def _populate_agent_tab(self):
-        self._load_agent_runtime_state()
-        preview = dict(getattr(self, "_agent_preview_override", {}) or {})
-        live_rows = list((self._agent_analysis or {}).get("stocks") or [])
-        if preview and not live_rows:
-            self._agent_analysis = preview
-        elif live_rows and not all(str(row.get("verdict") or "").upper() == "REVIEW" for row in live_rows):
-            self._agent_preview_override = None
-        a = getattr(self, '_agent_analysis', {})
-        self._update_agent_status("", "idle")
-        self._populate_agent_market_banner(a)
-        self._populate_agent_meta_headers(a)
-        self._render_agent_chat_history()
-
-        # Verdicts table — short summary, full reasoning on row select
-        dt = self.query_one("#dt-agent-verdicts", DataTable)
-        dt.clear(columns=True)
-        for label, key, width in [
-            (" #", "n", 2),
-            ("SYMBOL", "sym", 6),
-            ("SIGNAL", "algo", 9),
-            ("ACTION", "v", 8),
-            ("CONV", "conv", 4),
-            ("SCORE", "score", 5),
-            ("KEY POINT", "kp", 32),
-        ]:
-            dt.add_column(label, key=key, width=width)
-
-        stocks = list(a.get("stocks", []) or a.get("shortlist", []) or [])
-        if stocks:
-            for i, s in enumerate(stocks, 1):
-                verdict = s.get("verdict", "?").upper()
-                action_label = str(s.get("action_label") or verdict or "REVIEW").upper()
-                vc = {
-                    "BUY": f"bold {GAIN_HI}",
-                    "SELL": f"bold {LOSS_HI}",
-                    "PASS": f"bold {LOSS_HI}",
-                    "HOLD": f"bold {YELLOW}",
-                    "REVIEW": LABEL,
-                }.get(action_label, LABEL)
-                conv = s.get("conviction", 0)
-                conv_c = GAIN_HI if conv >= 0.7 else YELLOW if conv >= 0.4 else LOSS
-                # Extract first sentence as key point
-                reasoning = str(s.get("what_matters") or s.get("reasoning") or "")
-                first_sentence = reasoning.split(". ")[0].split(" — ")[0][:32]
-                score = float(s.get("signal_score") or 0.0)
-                score_style = GAIN_HI if score >= 1.0 else YELLOW if score >= 0.7 else LABEL
-                dt.add_row(
-                    _dim_text(f"{i:2d}"),
-                    _sym_text(s.get("symbol", "")),
-                    Text(str(s.get("signal_type") or s.get("algo_signal") or "")[:8], style=f"bold {AMBER}"),
-                    Text(f" {action_label} ", style=vc),
-                    Text(f"{conv:.0%}", style=conv_c),
-                    Text(f"{score:.2f}", style=score_style),
-                    Text(first_sentence, style=WHITE),
-                )
-        else:
-            dt.add_row(
-                _dim_text("—"), _dim_text("Loading top picks"),
-                *[Text("")] * 5)
-        self._populate_agent_detail_default(a)
-
     def on_data_table_row_selected(self, event: DataTable.RowSelected) -> None:
         """Handle row selection across tabs."""
         # Watchlist: Enter → lookup
@@ -2809,193 +2182,6 @@ class NepseDashboard(
         if event.data_table.id != "dt-agent-verdicts":
             return
         self._show_agent_focus_row(event.cursor_row)
-
-    def _append_chat(
-        self,
-        role: str,
-        message: str,
-        color: str,
-        *,
-        ts: float | None = None,
-        provider: str | None = None,
-    ):
-        scroll = self.query_one("#agent-chat-scroll", VerticalScroll)
-        self._detach_agent_typing_widget()
-        css_class = "chat-user" if role == "YOU" else "chat-agent"
-        label = "YOU" if role == "YOU" else "AGENT" if role == "AGENT" else role.upper()
-        ts_text = _format_nst_hm(ts)
-        flat_message = " ".join(part.strip() for part in str(message or "").splitlines() if part.strip())
-        if not flat_message:
-            flat_message = "—"
-        prefix = f"[{label}]"
-        msg_text = Text.assemble(
-            (prefix, f"bold {color}"),
-            (" ", WHITE),
-            (f"{ts_text} " if ts_text else "", LABEL),
-            (": ", LABEL),
-            (flat_message, WHITE),
-        )
-        widget = Static(msg_text, classes=f"chat-msg {css_class}")
-        scroll.mount(widget)
-        scroll.mount(Static("", classes="chat-gap"))
-        scroll.scroll_end(animate=False)
-        if bool(getattr(self, "_agent_typing_visible", False)):
-            self._animate_agent_typing()
-
-    def _handle_agent_chat_command(self, raw: str) -> bool:
-        command = str(raw or "").strip().lower()
-        if command in {"/history", "/archive", "/archived"}:
-            self._agent_show_archived = True
-            self._load_agent_runtime_state()
-            self._render_agent_chat_history()
-            self._set_status("Agent archive loaded")
-            return True
-        if command in {"/recent", "/hide", "/collapse"}:
-            self._agent_show_archived = False
-            self._load_agent_runtime_state()
-            self._render_agent_chat_history()
-            self._set_status("Agent archive hidden")
-            return True
-        if command == "/clear":
-            self._agent_show_archived = False
-            self._agent_visible_since = time.time()
-            self._load_agent_runtime_state()
-            self._render_agent_chat_history()
-            self._set_status("Agent chat screen cleared")
-            return True
-        if command == "/stop":
-            if self._stop_active_agent_chat():
-                return True
-            self._set_status("No active agent response to stop")
-            return True
-        if command in {"/agent", "/agent status"}:
-            summary = self._agent_runtime_summary()
-            self._append_chat_note(f"Active agent: {summary}. Config: {ACTIVE_AGENT_FILE}")
-            self._set_status(f"Active agent: {summary}")
-            return True
-        if command in {"/agent list", "/agents", "/backends"}:
-            help_text = self._agent_backends_help()
-            self._append_chat_note(help_text)
-            self._set_status(help_text)
-            return True
-        if command.startswith("/agent "):
-            try:
-                parts = shlex.split(raw)
-                if len(parts) < 2:
-                    raise ValueError("Usage: /agent <ollama|gemma4_mlx|gemma4_experimental|claude> [model]")
-                preset = parts[1].strip()
-                model = parts[2].strip() if len(parts) >= 3 else None
-                cfg = set_active_agent(preset, model=model)
-                summary = self._agent_runtime_summary()
-                self._append_chat_note(f"Agent switched: {summary}")
-                self._set_status(f"Agent switched to {cfg.get('selected_preset')} | model {cfg.get('model') or 'default'}")
-                self._populate_agent_tab()
-                return True
-            except Exception as exc:
-                self._set_status(f"Agent switch failed: {exc}")
-                return True
-        if command.startswith("/model"):
-            try:
-                parts = shlex.split(raw)
-                if len(parts) < 2:
-                    summary = self._agent_runtime_summary()
-                    self._append_chat_note(f"Active model: {summary}")
-                    self._set_status(f"Active agent: {summary}")
-                    return True
-                model = " ".join(parts[1:]).strip()
-                cfg = load_active_agent_config()
-                preset = str(cfg.get("selected_preset") or cfg.get("backend") or "ollama")
-                saved = set_active_agent(preset, model=model)
-                summary = self._agent_runtime_summary()
-                self._append_chat_note(f"Agent model updated: {summary}")
-                self._set_status(f"Agent model set to {saved.get('model')}")
-                self._populate_agent_tab()
-                return True
-            except Exception as exc:
-                self._set_status(f"Model update failed: {exc}")
-                return True
-        if command == "/help":
-            self._set_status("Agent commands: /agent, /agent list, /agent ollama gemma4:e2b, /model <name>, /history, /recent, /clear, /stop")
-            return True
-        return False
-
-    def _build_agent_auto_order_spec(self, analysis: dict) -> Optional[dict]:
-        if str(getattr(self, "trade_mode", "paper")) != "paper":
-            return None
-        if not bool((analysis or {}).get("trade_today")):
-            return None
-        stocks = sorted(
-            [dict(item) for item in list((analysis or {}).get("stocks") or []) if bool(item.get("auto_entry_candidate"))],
-            key=lambda item: (
-                -float(item.get("signal_score") or 0.0),
-                -float(item.get("conviction") or 0.0),
-            ),
-        )
-        if not stocks:
-            return None
-        positions = list((self._stats or {}).get("positions") or [])
-        held_symbols = {str(pos.get("sym") or "").upper() for pos in positions}
-        open_buy_symbols = {
-            str(order.get("symbol") or "").upper()
-            for order in list(getattr(self, "_paper_orders", []) or [])
-            if str(order.get("status") or "").upper() == "OPEN" and str(order.get("action") or "").upper() == "BUY"
-        }
-        max_positions = int(
-            LONG_TERM_CONFIG.get("regime_max_positions", {}).get(
-                str((analysis or {}).get("regime") or "").lower(),
-                LONG_TERM_CONFIG.get("max_positions", 5),
-            )
-        )
-        if len(held_symbols | open_buy_symbols) >= max_positions:
-            return None
-        prices = self.md.ltps() if hasattr(self, "md") else {}
-        cash = float((self._stats or {}).get("cash") or 0.0)
-        nav = float((self._stats or {}).get("nav") or cash or INITIAL_CAPITAL)
-        per_position_budget = min(nav / max_positions, cash * 0.95) if max_positions > 0 else 0.0
-        for stock in stocks:
-            symbol = str(stock.get("symbol") or "").upper()
-            if symbol in held_symbols or symbol in open_buy_symbols:
-                continue
-            price = float(prices.get(symbol) or stock.get("last_price") or 0.0)
-            if price <= 0:
-                continue
-            quantity = int(per_position_budget / price)
-            if quantity < 10:
-                continue
-            order_key = f"{analysis.get('context_date') or ''}:{symbol}:{int(float(analysis.get('timestamp') or 0))}"
-            if order_key == self._last_agent_auto_order_key:
-                return None
-            return {
-                "symbol": symbol,
-                "quantity": quantity,
-                "price": round(price, 2),
-                "order_key": order_key,
-                "signal_score": float(stock.get("signal_score") or 0.0),
-                "conviction": float(stock.get("conviction") or 0.0),
-            }
-        return None
-
-    def _maybe_submit_agent_super_signal(self, analysis: dict) -> None:
-        spec = self._build_agent_auto_order_spec(analysis or {})
-        if not spec:
-            return
-        result = build_tui_control_plane(self).submit_paper_order(
-            action="buy",
-            symbol=spec["symbol"],
-            quantity=int(spec["quantity"]),
-            limit_price=float(spec["price"]),
-            thesis="agent_super_signal",
-            confidence=float(spec["conviction"]),
-            source_signals=["agent_super_signal"],
-        )
-        if not result.ok:
-            return
-        self._last_agent_auto_order_key = str(spec["order_key"])
-        self._append_chat_note(
-            f"AUTO BUY queued {spec['symbol']} x{spec['quantity']} @ {spec['price']:.2f} "
-            f"(score {spec['signal_score']:.2f}, conv {spec['conviction']:.0%})"
-        )
-        self._set_status(f"Agent auto-buy queued {spec['symbol']} x{spec['quantity']} @ {spec['price']:.2f}")
 
     def on_button_pressed(self, event: Button.Pressed) -> None:
         """Handle button presses."""
@@ -3302,12 +2488,6 @@ class NepseDashboard(
         self.on_option_list_option_highlighted(event)
 
     # ── Header / Index / Status bars ──────────────────────────────────────────
-
-    def _display_nav_mode_tag(self) -> str:
-        if self._trading_engine:
-            phase = self._trading_engine.phase.upper().replace("_", " ")
-            return f"[bold #00cfff]AUTO[/] [dim]{phase}[/]"
-        return "[bold #00ff7f]PAPER[/]"
 
     @staticmethod
     def _summarize_tms_watchlist_error(exc: Exception, action: str = "sync") -> str:
@@ -3643,14 +2823,6 @@ class NepseDashboard(
         if not config:
             config = copy.deepcopy(LONG_TERM_CONFIG)
         return config
-
-    def _apply_active_strategy_runtime(self) -> None:
-        config = self._active_strategy_config()
-        self.signal_types = list(config.get("signal_types") or list(LONG_TERM_CONFIG.get("signal_types") or []))
-        self.max_positions = int(config.get("max_positions") or LONG_TERM_CONFIG.get("max_positions") or 0)
-        self.holding_days = int(config.get("holding_days") or LONG_TERM_CONFIG.get("holding_days") or 0)
-        self.sector_limit = float(config.get("sector_limit") or LONG_TERM_CONFIG.get("sector_limit") or 0.0)
-        self.use_regime_filter = bool(config.get("use_regime_filter", True))
 
     def _load_strategies_registry(self) -> None:
         strategy_registry.ensure_builtin_strategies()
@@ -4436,191 +3608,11 @@ class NepseDashboard(
         _save_accounts_registry(registry)
         self._paper_accounts = accounts
 
-    def _activate_paper_account(self, account_id: str) -> str:
-        target_id = str(account_id or "").strip()
-        if not target_id:
-            raise ValueError("Select an account to activate")
-        if target_id == self._current_account_id:
-            return f"{target_id} is already active"
-        self._persist_active_account_snapshot()
-        source_dir = _account_dir(target_id)
-        _blank_account_files(source_dir)
-        backup_dir = self._backup_profile_targets(list(ACTIVE_ACCOUNT_FILES.values()))
-        for name, active_path in ACTIVE_ACCOUNT_FILES.items():
-            source = source_dir / name
-            target = Path(active_path)
-            if source.exists():
-                shutil.copy2(source, target)
-        self._current_account_id = target_id
-        self._selected_account_id = target_id
-        # Switch active engine reference to the newly activated account
-        self._trading_engine = self._account_engines.get(target_id)
-        target_account = self._strategy_account_binding(target_id) or {}
-        self._selected_strategy_id = str(target_account.get("strategy_id") or self._selected_strategy_id)
-        self._apply_active_strategy_runtime()
-        self._signals_table_cache_key = ""
-        self._signals_table_cache_payload = None
-        self._sync_agent_account_context_env()
-        profile = _load_profile_config()
-        profile["current_account_id"] = target_id
-        _save_profile_config(profile)
-        self._paper_watchlist = _load_watchlist()
-        self._watchlist = list(self._paper_watchlist)
-        self._load_paper_orders()
-        self._populate_portfolio_and_risk()
-        self._populate_trades_full()
-        self._render_hedge_panel()
-        self._populate_orders_tab()
-        self._populate_watchlist()
-        self._populate_paper_profile_panel(self._stats)
-        self._populate_strategies_tab()
-        self._load_signals_async(force=True)
-        backup_note = f" | backup {backup_dir.name}" if backup_dir else ""
-        active_name = next((str(a.get("name") or target_id) for a in self._paper_accounts if str(a.get("id") or "") == target_id), target_id)
-        return f"Activated {active_name}{backup_note}"
-
-    def _create_paper_account(self, raw_name: str, raw_portfolio_path: str, raw_target_nav: str) -> str:
-        name = str(raw_name or "").strip()
-        if not name:
-            raise ValueError("Enter an account name")
-        portfolio_path = _coerce_dragdrop_path(raw_portfolio_path)
-        token = str(raw_target_nav or "").strip().replace(",", "")
-        if not token:
-            raise ValueError("Enter a target NAV")
-        target_nav = float(token)
-        if target_nav <= 0:
-            raise ValueError("Target NAV must be positive")
-        if portfolio_path:
-            if not portfolio_path.exists():
-                raise ValueError("Portfolio CSV path not found")
-            df = pd.read_csv(portfolio_path)
-            portfolio_df = _normalize_import_portfolio(df)
-        else:
-            portfolio_df = pd.DataFrame(columns=PORTFOLIO_COLS)
-        state, nav_log = _build_account_seed_state(portfolio_df, target_nav)
-        registry = _load_accounts_registry()
-        accounts = list(registry.get("accounts") or [])
-        if len(accounts) >= MAX_ACCOUNTS:
-            raise ValueError(f"Maximum {MAX_ACCOUNTS} accounts allowed")
-        if any(str(account.get("name") or "").strip().lower() == name.lower() for account in accounts):
-            raise ValueError(f"{name} already exists")
-        account_id = _next_account_id(accounts)
-        target_dir = _account_dir(account_id)
-        _blank_account_files(target_dir)
-        portfolio_df.reindex(columns=PORTFOLIO_COLS).to_csv(target_dir / "paper_portfolio.csv", index=False)
-        portfolio_df.reindex(columns=PORTFOLIO_COLS).to_csv(target_dir / "tui_paper_portfolio.csv", index=False)
-        pd.DataFrame(columns=TRADE_LOG_COLS).to_csv(target_dir / "paper_trade_log.csv", index=False)
-        pd.DataFrame(columns=TRADE_LOG_COLS).to_csv(target_dir / "tui_paper_trade_log.csv", index=False)
-        nav_log.reindex(columns=NAV_LOG_COLS).to_csv(target_dir / "paper_nav_log.csv", index=False)
-        nav_log.reindex(columns=NAV_LOG_COLS).to_csv(target_dir / "tui_paper_nav_log.csv", index=False)
-        save_runtime_state(str(target_dir / "paper_state.json"), dict(state))
-        save_runtime_state(str(target_dir / "tui_paper_state.json"), dict(state))
-        watchlist_entries = _build_holdings_watchlist_entries(portfolio_df)
-        (target_dir / "watchlist.json").write_text(_json.dumps(_merge_watchlist_entries(watchlist_entries), indent=2))
-        (target_dir / "tui_paper_orders.json").write_text("[]")
-        (target_dir / "tui_paper_order_history.json").write_text("[]")
-        now_stamp = datetime.now().isoformat(timespec="seconds")
-        accounts.append({
-            "id": account_id,
-            "name": name,
-            "strategy_id": strategy_registry.default_strategy_for_account(account_id),
-            "created_at": now_stamp,
-            "updated_at": now_stamp,
-        })
-        registry["accounts"] = accounts
-        _save_accounts_registry(registry)
-        self._paper_accounts = strategy_registry.ensure_account_strategy_ids(accounts)
-        self._persist_profile_paths(portfolio_path=str(portfolio_path) if portfolio_path else "", target_nav=target_nav)
-        self.query_one("#profile-inp-account-name", Input).value = f"Account {len(accounts) + 1}"
-        self._selected_account_id = account_id
-        # Start engine for the new account if multi-engine is running
-        if self.trade_mode == "paper" and account_id not in getattr(self, "_account_engines", {}):
-            _new_strategy = strategy_registry.load_strategy(strategy_registry.default_strategy_for_account(account_id))
-            _new_config = dict((_new_strategy or {}).get("config") or {}) or dict(LONG_TERM_CONFIG)
-            _new_adir = _account_dir(account_id)
-            _new_capital = _account_initial_capital_from_files(
-                _new_adir,
-                float(_new_config.get("initial_capital") or target_nav or INITIAL_CAPITAL),
-            )
-            _new_engine = TUITradingEngine(
-                capital=_new_capital,
-                signal_types=list(_new_config.get("signal_types") or list(LONG_TERM_CONFIG.get("signal_types") or [])),
-                max_positions=int(_new_config.get("max_positions") or LONG_TERM_CONFIG.get("max_positions") or 5),
-                holding_days=int(_new_config.get("holding_days") or LONG_TERM_CONFIG.get("holding_days") or 40),
-                sector_limit=float(_new_config.get("sector_limit") or LONG_TERM_CONFIG.get("sector_limit") or 0.35),
-                stop_loss_pct=float(_new_config.get("stop_loss_pct") or LONG_TERM_CONFIG.get("stop_loss_pct") or 0.08),
-                trailing_stop_pct=float(_new_config.get("trailing_stop_pct") or LONG_TERM_CONFIG.get("trailing_stop_pct") or 0.10),
-                hedge_enabled=bool(getattr(self, "_hedge_enabled", True)),
-                portfolio_file=_new_adir / "paper_portfolio.csv",
-                trade_log_file=_new_adir / "paper_trade_log.csv",
-                nav_log_file=_new_adir / "paper_nav_log.csv",
-                state_file=_new_adir / "paper_state.json",
-                account_id=account_id,
-                strategy_id=strategy_registry.default_strategy_for_account(account_id),
-                strategy_config=_new_config,
-                on_status=self._make_engine_status_cb(account_id),
-                on_activity=self._make_engine_activity_cb(account_id),
-                on_portfolio_changed=self._make_engine_portfolio_cb(account_id),
-                on_agent_updated=self._make_engine_agent_cb(account_id),
-            )
-            self._account_engines[account_id] = _new_engine
-            self._start_account_engine_worker(account_id)
-        message = self._activate_paper_account(account_id)
-        seed_note = f"{portfolio_df.shape[0]} seeded holdings" if not portfolio_df.empty else "cash-only account"
-        return f"Created {name} | {seed_note} | {message}"
-
     def _save_current_account_snapshot(self) -> str:
         self._persist_active_account_snapshot()
         self._populate_paper_profile_panel(self._stats)
         active_name = next((str(a.get("name") or self._current_account_id) for a in self._paper_accounts if str(a.get("id") or "") == self._current_account_id), self._current_account_id)
         return f"Saved snapshot for {active_name}"
-
-    def _delete_paper_account(self, account_id: str) -> str:
-        """Stop engine, backup files, remove account from registry."""
-        target_id = str(account_id or "").strip()
-        if not target_id:
-            raise ValueError("No account selected")
-        registry = _load_accounts_registry()
-        accounts = list(registry.get("accounts") or [])
-        if len(accounts) <= 1:
-            raise ValueError("Cannot delete the only account")
-        target = next((a for a in accounts if str(a.get("id") or "") == target_id), None)
-        if not target:
-            raise ValueError(f"Account {target_id} not found")
-        # Stop and remove the account's engine
-        engine = self._account_engines.pop(target_id, None)
-        if engine:
-            try:
-                engine.stop()
-            except Exception:
-                pass
-        if self._trading_engine is engine:
-            self._trading_engine = None
-        # Backup account directory before deleting
-        account_dir = _account_dir(target_id)
-        if account_dir.exists():
-            try:
-                stamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-                backup = ensure_dir(PAPER_IMPORT_BACKUP_DIR / f"deleted_{target_id}_{stamp}")
-                shutil.copytree(account_dir, backup / target_id)
-                shutil.rmtree(account_dir, ignore_errors=True)
-            except Exception:
-                pass
-        # Remove from registry
-        remaining = [a for a in accounts if str(a.get("id") or "") != target_id]
-        registry["accounts"] = remaining
-        _save_accounts_registry(registry)
-        self._paper_accounts = strategy_registry.ensure_account_strategy_ids(remaining)
-        name = str(target.get("name") or target_id)
-        # If deleting the active account, activate the first remaining one
-        if target_id == self._current_account_id and remaining:
-            try:
-                self._activate_paper_account(str(remaining[0].get("id") or ""))
-            except Exception:
-                pass
-        else:
-            self._populate_account_list()
-        return f"Deleted '{name}' (backup saved)"
 
     def _copy_import_dataframe(
         self,
